ai_agent/requirements.txt
openai-agents
docker